soilwise-app/requirements.txt
streamlit
pydub
SpeechRecognition
pillow
openai
googletrans==4.0.0-rc1
